refactor(installer): extract describePlan + tighten copilot test
