+++ scripts/installer/targets/claude.ts
@@ -5,6 +5,7 @@ import {
   checkTargetDirWritable,
   checkDockerWarn,
 } from "../preflight.js";
+import { describePlan } from "./shared.js";
 
 const HOOK_SCRIPTS = [
   "memory-session-start.sh",
@@ -70,12 +71,5 @@ export const claudeTarget: Target = {
     };
   },
 
-  describe(plan: InstallPlan): string {
-    const lines: string[] = [`Target: ${plan.target}`];
-    for (const c of plan.copies) lines.push(`  copy  ${c.src} → ${c.dest}`);
-    for (const m of plan.jsonMerges) lines.push(`  merge ${m.file}`);
-    for (const p of plan.markdownPrepends)
-      lines.push(`  prepend ${p.file} [${p.markerId}]`);
-    return lines.join("\n");
-  },
+  describe: describePlan,
 };
